Squashed commit of the following:

commit 79c6522e64729d4727ab42259e9c77273bde4a22
Author: DEXREQ Automation <noreply+dexreq@oracle.com>
Date:   Fri Jan 24 19:12:04 2020 +0000

    Author: DEXREQ Automation <noreply+dexreq@oracle.com>
    Date:  Fri, 24 Jan 2020 19:12:04 +0000
    Running generation for: [[DEXREQ-997, DEXREQ-994, DEXREQ-1000, DEXREQ-990, DEXREQ-998, DEXREQ-1001, DEXREQ-995]]

    Author: DEXREQ Automation <noreply+dexreq@oracle.com>
    Date:  Fri, 24 Jan 2020 18:46:27 +0000
    Updating pom.xml for [[DEXREQ-997, DEXREQ-994, DEXREQ-1000, DEXREQ-990, DEXREQ-998, DEXREQ-1001, DEXREQ-995]]

    (automatically combined)

+++ core/image_ocpu_constraints.go
@@ -20,10 +20,10 @@ import (
 type ImageOcpuConstraints struct {
 
 	// The minimum number of OCPUs supported for this image and shape.
-	Min *float32 `mandatory:"false" json:"min"`
+	Min *int `mandatory:"false" json:"min"`
 
 	// The maximum number of OCPUs supported for this image and shape.
-	Max *float32 `mandatory:"false" json:"max"`
+	Max *int `mandatory:"false" json:"max"`
 }
 
 func (m ImageOcpuConstraints) String() string {
